refactor: 패키지를 game으로 이동
