+++ src/test/java/baseball/domain/game/StartedTest.java
@@ -1,69 +1,56 @@
-package baseball.domain;
+package baseball.domain.game;
 
 import static baseball.domain.digits.DigitsUtil.digitsOf;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
-import static org.assertj.core.api.Assertions.assertThatIllegalStateException;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import baseball.domain.digits.Digits;
-import baseball.domain.digits.GuessAnswer;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-class IncorrectTest {
+class StartedTest {
     private static final Digits SECRET_NUMBER = digitsOf(1, 2, 3);
-    private static final GuessAnswer PREV_GUESS_ANSWER = new GuessAnswer(3, 1, 2);
-    private final GameState incorrect = new Incorrect(SECRET_NUMBER, PREV_GUESS_ANSWER);
+    private final GameState started = new Started(SECRET_NUMBER);
 
-    @DisplayName("비밀번호와 이전 추측결과로 생성된다.")
+    @DisplayName("비밀번호로 생성된다.")
     @Test
     void construct() {
         assertThatNoException()
-                .isThrownBy(() -> new Incorrect(SECRET_NUMBER, PREV_GUESS_ANSWER));
+                .isThrownBy(() -> new Started(SECRET_NUMBER));
     }
 
-    @DisplayName("이전 추측결과 맞추지 못했어야 한다.")
-    @Test
-    void illegalGuessAnswer() {
-        final GuessAnswer prevGuessAnswer = new GuessAnswer(3, 3, 0);
-
-        assertThatIllegalStateException()
-                .isThrownBy(() -> new Incorrect(SECRET_NUMBER, prevGuessAnswer))
-                .withMessage("Previous guess answer is correct!");
-    }
-
-    @DisplayName("맞추지 못했다.")
+    @DisplayName("시작했기에 아직 맞추지 못했다.")
     @Test
     void correct() {
-        assertThat(incorrect.isCorrect()).isFalse();
+        assertThat(started.isCorrect()).isFalse();
     }
 
-    @DisplayName("스트라이크 개수는 이전 추측결과의 개수이다.")
+    @DisplayName("시작했기에 스트라이크 개수는 0이다.")
     @Test
     void countStrikes() {
-        assertThat(incorrect.countStrikes()).isEqualTo(1);
+        assertThat(started.countStrikes()).isZero();
     }
 
-    @DisplayName("볼 개수는 이전 추측결과의 개수이다.")
+    @DisplayName("시작했기에 볼 개수는 0이다.")
     @Test
     void countBalls() {
-        assertThat(incorrect.countBalls()).isEqualTo(2);
+        assertThat(started.countBalls()).isZero();
     }
 
-    @DisplayName("낫싱은 이전 추측결과로 정해진다.")
+    @DisplayName("시작했기에 낫싱이다.")
     @Test
     void isNothing() {
-        assertThat(incorrect.isNothing()).isEqualTo(false);
+        assertThat(started.isNothing()).isTrue();
     }
 
     @DisplayName("종료되지 않았다.")
     @Test
     void isFinished() {
-        assertThat(incorrect.isFinished()).isFalse();
+        assertThat(started.isFinished()).isFalse();
     }
 
     @DisplayName("비밀번호와 추측한 숫자들의 길이가 같아야 한다.")
@@ -72,7 +59,7 @@ class IncorrectTest {
         final Digits guessDigits = digitsOf(1, 2, 3, 4);
 
         assertThatIllegalArgumentException()
-                .isThrownBy(() -> incorrect.guess(guessDigits))
+                .isThrownBy(() -> started.guess(guessDigits))
                 .withMessage("Guess digits size must be equal to secret number size!");
 
     }
@@ -95,7 +82,7 @@ class IncorrectTest {
     ) {
         final Digits guessDigits = digitsOf(guess1st, guess2nd, guess3rd);
 
-        final GameState nextState = incorrect.guess(guessDigits);
+        final GameState nextState = started.guess(guessDigits);
 
         assertAll(
                 () -> assertThat(nextState.countStrikes()).isEqualTo(expectedStrikes),
